+++ packages/web/src/lib/wallectconnect-mode.ts
@@ -162,13 +162,14 @@ export class WallectConnectSession {
 	 * @param transaction algosdk.Transaction object
 	 * @param logicSig Logic Sig Account
 	 * @returns Returns txID and blob object
+	 * for more info: https://developer.algorand.org/docs/get-details/dapps/smart-contracts/smartsigs/modes/#contract-account
 	 */
-	signLogicSignatureTxn(transaction: Transaction, logicSig: algosdk.LogicSigAccount): { txID: string, blob: Uint8Array } {
+	signLogicSigTx(transaction: Transaction, logicSig: algosdk.LogicSigAccount): { txID: string, blob: Uint8Array } {
 		try {
 			return algosdk.signLogicSigTransaction(transaction, logicSig)
 		} catch (err) {
 			error(err);
-			throw new Error("Error while signing Lsig Transaction" + err);
+			throw err
 		}
 	}
 
@@ -315,7 +316,7 @@ export class WallectConnectSession {
 				if (signer.sign === SignType.LogicSignature) {
 					signer.lsig.lsig.args = signer.args ? signer.args : [];
 					if (!Array.isArray(signedTxn)) signedTxn = [];
-					signedTxn.splice(index, 0, this.signLogicSignatureTxn(txn, signer.lsig).blob);
+					signedTxn.splice(index, 0, this.signLogicSigTx(txn, signer.lsig).blob);
 				}
 			}