+++ packages/web/test/src/lib/myalgowallet-mode.ts
@@ -128,7 +128,7 @@ describe("Webmode - MyAlgo Wallet test cases ", () => {
 			const txnParams = await getSuggestedParams(algodClient);
 			const transactions: Transaction[] = connector.makeTx([execParams], txnParams);
 			assert.doesNotThrow(async () => {
-				const signTx = connector.signLogicSignatureTxn(transactions[0], lsigAccount);
+				const signTx = connector.signLogicSigTx(transactions[0], lsigAccount);
 				expect(signTx).to.have.ownProperty("blob");
 				expect(signTx).to.have.ownProperty("txID");
 				expect(signTx.blob).to.exist;